Move patches into separate files
