+++ client-side/InventoryOrganizingFeatures/patches/PreItemViewOnBeginDrag.cs
@@ -1,0 +1,53 @@
+using Aki.Reflection.Patching;
+using EFT.HandBook;
+using EFT.InventoryLogic;
+using EFT.UI;
+using EFT.UI.DragAndDrop;
+using HarmonyLib;
+using InventoryOrganizingFeatures.Reflections;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Xml.Linq;
+using UnityEngine;
+using UnityEngine.UI;
+using static InventoryOrganizingFeatures.Locker;
+using static InventoryOrganizingFeatures.Organizer;
+using static InventoryOrganizingFeatures.OrganizedContainer;
+using static InventoryOrganizingFeatures.UserInterfaceElements;
+using InventoryOrganizingFeatures.Reflections.Extensions;
+using TMPro;
+using BepInEx.Logging;
+using Debug = UnityEngine.Debug;
+
+namespace InventoryOrganizingFeatures
+{
+    internal class PreItemViewOnBeginDrag : ModulePatch
+    {
+        protected override MethodBase GetTargetMethod()
+        {
+            return AccessTools.Method(typeof(ItemView), "OnBeginDrag");
+        }
+
+        [PatchPrefix]
+        private static bool PatchPrefix(ref ItemView __instance)
+        {
+            try
+            {
+                // Don't execute original event handler if item IsMoveLocked, otherwise execute.
+                if (IsMoveLocked(__instance.Item)) return false;
+                return true;
+            }
+            catch (Exception ex)
+            {
+                throw Plugin.ShowErrorNotif(ex);
+            }
+        }
+    }
+}
